Guarantee consistent test results
We do this by asking Monument for more compositions than are possible to
generate in the search (so, for example, if the search would produce 324
comps, we would ask it for 400).  The tests are constructed to be small
enough that Monument never runs out of queue space, so Monument _must_
generate exactly the tests given.  If/when Monument can provide hard
guarantees about optimality (i.e. it will guarantee to only output _the
best_ comps) then the test results will be consistent anyway and this
restriction can be removed.