Return owned data from `MethodLib::get_by_title_with_suggestions`
Previously, `MethodLib::get_by_title_with_suggestions` would return a
reference to method titles within the `MethodLib`.  This avoids cloning
a few `String`s but makes for a nastier API if the error is expected to
outlive the `MethodLib` it came from (which happens in Monument).
Cloning a few small `String`s is never going to be a bottleneck - and
we're not even in the happy path anymore because we're handling errors.